Call isAuthenticated directly in views
This is the easiest way to keep the view up-to-date with the
authentication status, as the intention is that the status should be
correct every time the page is rendered. This will lead to the overhead
of reading from local storage multiple times, but that could be
mitigated by saving the token data in RAM if it proves to be a problem.