Use table instead of list in match-list
To be able to present the information in a more readable way, this
is changed to be in a table instead of a list. The information shown
has also been simplified.

Will later change so that this table is sorted on date, when that is
available in the match API