Show session-expired message for user on 401
The user now is prompted with the login screen once a 401 response is
received from the API. Previously this would not happen as the user
credentials (including token) was still in storage, which meant that the
user was rerouted to `/home` at all times.

Furthermore, the user is also given a reason for why the login screen is
prompted again (session expired).

403 errors are no longer handled by the authentication interceptor, as
there cannot be a standard way to handle them. No new authentication
would help the client.