Remove excessive authentication checks
The HTML templates have been doing a lot of function calls to do
authentication checks. This is very demanding on the performance of the
app and should be avoided.

To avoid this `AuthenticatedService.isAuthenticated` was remade to
return an `Observable` instead. This means that interested components
can subscribe and be automatically be notified when a change in the
authentication status has occurred, and consequently they do not have to
poll for changes for every render.